landingpage+login

+++ database/migrations/2024_10_12_081535_create_orders_table.php
@@ -1,29 +1,28 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Support\Facades\DB; // إضافة هذه السطر
-
 
 return new class extends Migration
 {
     /**
      * Run the migrations.
      */
-    public function up()
-    {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamp('order_date')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->enum('status', ['pending', 'shipped', 'completed']);
-            $table->decimal('total_amount', 10, 2);
-            $table->timestamps();
-        });
-    }
-    
 
+    public function up()
+        {
+            if (!Schema::hasTable('orders')) {
+                Schema::create('orders', function (Blueprint $table) {
+                    $table->id();
+                    $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                    $table->timestamp('order_date')->useCurrent(); 
+                    $table->enum('status', ['pending', 'shipped', 'completed']); 
+                    $table->decimal('total_amount', 10, 2)->nullable(); 
+                    $table->timestamps();
+                });
+            }
+        }
+        
 
     /**
      * Reverse the migrations.
